fix(core): Caps auf wasm-Speichergrenzen bringen, Manifest-Abgleich und Testluecken schliessen
Groessen-Caps waren fuer den tatsaechlichen wasm32-Zieltyp zu hoch gewaehlt
und haetten dort selbst einen Abort statt eines CoreError ausgeloest; der
Schema-Version-Abgleich im Manifest hat bei jeder migrierten Datei falsch
Alarm geschlagen. Hash- und Manifest-Abgleich sind jetzt direkt getestet
statt nur behauptet.

+++ crates/videola-core/src/error.rs
@@ -40,5 +40,5 @@ pub enum CoreError {
     NotAProject(String),
 
     #[error("unsupported schema version {0}")]
-    UnsupportedSchema(u32),
+    UnsupportedSchema(u64),
 }
